+++ R/utilities.R
@@ -52,6 +52,12 @@ dt_keep_cols <- function(DT, ...){
   if(!("data.table" %in% class(DT)))
     stop("DT must be a data.table object")
   cols <- c(...)
+  badp <- setdiff(cols, names(DT))
+  if(length(badp)){
+    missingnames <- paste(sprintf('"%s"', badp), collapse=', ')
+    warning(sprintf("Columns (%s) are missing from supplied data.table",
+                    missingnames))
+  }
   these <- setdiff(names(DT), cols)
   dt_del_cols(DT, these)
 }
@@ -120,6 +126,10 @@ dt_pivot <- function(DT, theby, theexp, percent.cutoff=0, value.name="value",
 #'        the count percents lower than this number will be
 #'        grouped into "OTHER" in the returned data.table
 #'        (default is 0)
+#' @param big.mark If \code{FALSE} (default) the "count" column is left
+#'                 as an integer. If not \code{FALSE}, it must be a
+#'                 character to separate every three digits of the count.
+#'                 This turns the count column into a string.
 #'
 #' @return Returns a data.table with three columns:
 #'         the grouped-by column, a count column, and a
@@ -136,12 +146,23 @@ dt_pivot <- function(DT, theby, theexp, percent.cutoff=0, value.name="value",
 #' dt_counts_and_percents(mt, "cyl", percent.cutoff=25)
 #'
 #' @export
-dt_counts_and_percents <- function(DT, group_by_this, percent.cutoff=0){
+dt_counts_and_percents <- function(DT, group_by_this, percent.cutoff=0,
+                                   big.mark=FALSE){
   if(!("data.table" %in% class(DT)))
     stop("DT must be a data.table object")
-  dt_pivot(DT, group_by_this, .N, percent.cutoff=percent.cutoff,
-           value.name="count", percent.name="percent")
+  if(!(big.mark==FALSE || is.character(big.mark)))
+    stop("big.mark must either be FALSE or a character type")
+
+  count <- NULL
+
+  tmp <- dt_pivot(DT, group_by_this, .N, percent.cutoff=percent.cutoff,
+                  value.name="count", percent.name="percent")
+  if(big.mark!=FALSE)
+    tmp[, count:=prettyNum(count, big.mark=big.mark)]
+  tmp[]
 }
+
+
 
 
 # --------------------------------------------------------------- #
@@ -155,29 +176,46 @@ dt_counts_and_percents <- function(DT, group_by_this, percent.cutoff=0){
 #' All space/whitespace characters are replaced with underscores,
 #' as are all characters not from A-Z, a-z, an underscore, or a digit
 #'
+#' If there are duplicate column names after the cleaning, a message
+#' will show stating such and the duplicate column names will be
+#' make unique.
+#'
 #' @param dat A data.frame
+#' @param lower A logical indicating whether all column names should be
+#'              made lower case (default \code{TRUE}).
 #'
 #' @return Returns a vector of cleaned names
+#'
+#' @seealso \code{\link{make.unique}}
 #'
 #' @examples
 #' ejemplo <- iris
 #' names(ejemplo) <- c("Sepal Length", "Sepal@Width", "Petal	Length",
 #'                     "Petal\\nWidth", "Spêcies")
 # get_clean_names(ejemplo)
+#' # c("sepal_length" "sepal_width"  "petallength"  "petal_nwidth" "sp_cies")
+#'
+# get_clean_names(ejemplo, lower=FALSE)
 #' # c("Sepal_Length" "Sepal_Width"  "PetalLength"  "Petal_nWidth" "Sp_cies")
 #'
 #' @export
-get_clean_names <- function(dat){
+get_clean_names <- function(dat, lower=TRUE){
   thenames <- names(dat)
   nospace <- stringr::str_replace_all(thenames, "\\s+", "_")
   nobad <- stringr::str_replace_all(nospace, "[^_A-Za-z1-9]", "_")
+  if(lower)
+    nobad <- tolower(nobad)
+  if(sum(duplicated(nobad))){
+    message("cleaned names contain duplicates... correcting")
+    nobad <- make.unique(nobad, sep="_")
+  }
   return(nobad)
 }
 
 
 # --------------------------------------------------------------- #
 
-#' Takes a data.table and returns
+#' Takes a data.table and set to cleaned column names
 #'
 #' This function takes a data.table, and returns the same data.table
 #' with column names that are cleaned and stripped of potentially
@@ -189,6 +227,8 @@ get_clean_names <- function(dat){
 #' @import data.table
 #'
 #' @param DT a data.table
+#' @param lower A logical indicating whether all column names should be
+#'              made lower case (default \code{TRUE}).
 #'
 #' @return Returns the data.table but with cleaned names
 #'
@@ -202,10 +242,10 @@ get_clean_names <- function(dat){
 #' dt_set_clean_names(ejemplo)
 #'
 #' @export
-dt_set_clean_names <- function(DT){
+dt_set_clean_names <- function(DT, lower=TRUE){
   if(!("data.table" %in% class(DT)))
     stop("DT must be a data.table object")
-  setnames(DT, get_clean_names(DT))
+  setnames(DT, get_clean_names(DT, lower=lower))
 }
 
 
@@ -241,6 +281,57 @@ dt_percent_not_na <- function(DT, acolumn){
 }
 
 
+
+# --------------------------------------------------------------- #
+
+#' Get a breakdown of the NA-status of a column in a data.table
+#'
+#' This function takes a (quoted) column to group by, and tabulates
+#' the count of how many of those values are not-NA and NA, and adds
+#' the percent of occurrences. A \code{TRUE} in the first output
+#' column means the data is _not_ missing; \code{FALSE} corresponds
+#' to missing.
+#'
+#' The final row is a total count
+#'
+#' The quoted group-by variable must be a character or factor
+#'
+#' @import data.table
+#'
+#' @param DT The data.table object to operate on
+#' @param acolumn a quoted column name
+#' @param big.mark If \code{FALSE} (default) the "count" column is left
+#'                 as an integer. If not \code{FALSE}, it must be a
+#'                 character to separate every three digits of the count.
+#'                 This turns the count column into a string.
+#'
+#' @return Returns a data.table with three columns:
+#'         the not-NA status of the column specified, a count column, and a
+#'         percent column (out of 100) to two decimal places
+#'
+#' @examples
+#'
+#' iris_dt <- as.data.table(iris)
+#' iris_dt[sample(1:.N, 10), Species:=NA_character_]
+#' dt_na_breakdown(iris_dt, "Species")
+#'
+#' @export
+dt_na_breakdown <- function(DT, acolumn, big.mark=FALSE){
+  if(!("data.table" %in% class(DT)))
+    stop("DT must be a data.table object")
+  if(!(big.mark==FALSE || is.character(big.mark)))
+    stop("big.mark must either be FALSE or a character type")
+
+  count <- . <- NULL
+
+  tmp <- dt_pivot(DT[, .(not_na=as.character(!is.na(get(acolumn))))],
+            "not_na", .N, value.name="count", percent.name="percent")
+  if(big.mark!=FALSE)
+    tmp[, count:=prettyNum(count, big.mark=big.mark)]
+  setnames(tmp, "not_na", sprintf("%s_not_na", acolumn))
+  tmp[]
+}
+
 # --------------------------------------------------------------- #
 
 #' Add string to all column names in a data.table
@@ -260,8 +351,18 @@ dt_percent_not_na <- function(DT, acolumn){
 #'        Cannot co-exist with \code{include}
 #' @param include A quoted vector or column names. Changes names of only
 #'        these columns. Cannot co-exist with \code{exclude}
+#' @param fix.duplicates A logical indicating whether to, if after the
+#'                       suffix/prefixes are added to the selected column
+#'                       names, correct those duplicate column names by
+#'                       making them unique. If \code{FALSE} (default),
+#'                       if any of the column names are duplicated, an
+#'                       error is raised and the new names are not set.
+#'                       If \code{TRUE}, all the column names are made unique,
+#'                       potentially renaming excluded column names that
+#'                       were not supposed to be changed.
 #'
-#' @return Returns data.table with string appended or prefixed
+#' @return Returns data.table with string appended or prefixed to all selected
+#'         column names.
 #'
 #' @examples
 #' DT <- as.data.table(iris)
@@ -285,7 +386,8 @@ dt_percent_not_na <- function(DT, acolumn){
 #'
 #' @export
 dt_add_to_col_names <- function(DT, astring, prefix=FALSE,
-                                exclude=NULL, include=NULL){
+                                exclude=NULL, include=NULL,
+                                fix.duplicates=FALSE){
   if(!("data.table" %chin% class(DT)))
     stop("DT must be a data.table")
   if(!is.null(exclude) && !is.null(include))
@@ -315,5 +417,72 @@ dt_add_to_col_names <- function(DT, astring, prefix=FALSE,
   if(!is.null(include))
     tmp[!(old_names %chin% include), new_names:=old_names]
 
+  if(tmp[,sum(duplicated(new_names))]){
+    if(fix.duplicates){
+      message("cleaned names contain duplicates... correcting")
+      tmp[, new_names:=make.unique(new_names, sep="_")]
+    }
+    else
+      stop("not resetting column names... new names contain duplicates")
+  }
+
   setnames(DT, tmp[,new_names])
 }
+
+
+# --------------------------------------------------------------- #
+
+# unexported function to assist with WorldCat Search API SRU queries
+sru_syntax_translate_worldcat <- function(x){
+  index.xlation.table <- list(
+    c("srw.kw",  "\\$keyword"),
+    c("srw.ti",  "\\$title"),
+    c("srw.ln",  "\\$language"),
+    c("srw.au",  "\\$author"),
+    c("srw.yr",  "\\$year"),
+    c("srw.su",  "\\$subject"),
+    c("srw.li",  "\\$holding_library"),
+    c("srw.mt",  "\\$material_type"),
+    c("srw.no",  "\\$oclc"),
+    c("srw.lc",  "\\$lc_call"),
+    c("srw.dd",  "\\$dewey"),
+    c("srw.dn",  "\\$lccn"),
+    c("srw.bn",  "\\$isbn"),
+    c("srw.in",  "\\$issn"),
+    c("srw.cg",  "\\$library_holdings_group"),
+    c("srw.la",  "\\$language_code"),
+    c("srw.pl",  "\\$place_of_publication"),
+    c("srw.pb",  "\\$publisher"),
+    c("srw.am",  "\\$access_method"),
+    c("srw.cn",  "\\$corporate_conference_name"),
+    c("srw.pc",  "\\$dlc_limit"),
+    c("srw.dt",  "\\$document_type"),
+    c("srw.gn",  "\\$government_document_number"),
+    c("srw.mn",  "\\$music_publisher_number"),
+    c("srw.nt",  "\\$notes"),
+    c("srw.on",  "\\$open_digital_limit"),
+    c("srw.pn",  "\\$personal_name"),
+    c("srw.se",  "\\$series"),
+    c("srw.sn",  "\\$standard_number")
+  )
+
+  xlate.each.index.one <- function(y){
+    x <<- stringr::str_replace_all(x,
+                                   # negative lookbehind assertion
+                                   sprintf('(?<![\\w\\\\])%s', y[2]),
+                                   y[1])
+  }
+
+  rem.escapes <- function(y){
+    x <<- gsub(sprintf('%s', y[2]), gsub("^.", "", y[2], perl=TRUE), x, fixed=TRUE)
+  }
+
+  lapply(index.xlation.table, xlate.each.index.one)
+  lapply(index.xlation.table, rem.escapes)
+
+  # mat type
+
+  x
+}
+
+